Guess this fixed video notes

+++ main.py
@@ -75,8 +75,7 @@ def handle(msg):
         else:
             os.system("ffmpeg -ss 0 -t 60 -y -i " + filename + " -strict -2 -ac 1 -map 0:a -codec:a opus -b:a 128k -vn output.ogg")
         sendVoice(chat_id, "output.ogg")
-        #TODO: Making encoding video notes properly working, so that it will be displayed correctly in the client
-    '''if content_type == "video":
+    if content_type == "video":
         os.system("rm -f *.mp4")
         videofile = msg['video']
         fileid = msg['video']['file_id']
@@ -91,12 +90,10 @@ def handle(msg):
         length = video.duration * 0.33
         l2 = (video.duration * 0.33) + 60
         if video.duration > l2:
-            os.system("ffmpeg -ss " + str(length) + " -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=640:-1 vm.mp4")
+            os.system("ffmpeg -ss " + str(length) + " -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=480:480 vm.mp4")
         else:
-            os.system("ffmpeg -ss 0 -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=640:-1 vm.mp4")
-        f = open("vm.mp4", "r")
-        bot.sendVideoNote(chat_id, f)
-        f.close()'''
+            os.system("ffmpeg -ss 0 -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=480:480 vm.mp4")
+        sendVideoNote("vm.mp4")
     if content_type == "text":
         os.system("rm -f audio.jpg")
         if msg['text'].startswith("/chatid"):
@@ -269,18 +266,15 @@ def handle(msg):
                 cmd_conv = "ffmpeg -y -i video.mp4 -c:v libx264 -crf 26 -vf scale=640:-1 -strict -2 out.mp4"
                 bot.editMessageText(msgid, "Converting...")
                 subprocess.Popen(cmd_conv.split(), shell=False).wait()
-                #TODO: Making encoding video notes properly working, so that it will be displayed correctly in the client
-                '''filename = "out.mp4"
+                filename = "out.mp4"
                 video = VideoFileClip(filename)
                 length = video.duration * 0.33
                 l2 = (video.duration * 0.33) + 60
                 if video.duration > l2:
-                    os.system("ffmpeg -ss " + str(length) + " -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=640:-1 vm.mp4")
+                    os.system("ffmpeg -ss " + str(length) + " -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=480:480 vm.mp4")
                 else:
-                    os.system("ffmpeg -ss 0 -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=640:-1 vm.mp4")
-                f = open("vm.mp4", "r")
-                bot.sendVideoNote(chat_id, f)
-                f.close()'''
+                    os.system("ffmpeg -ss 0 -t 59 -y -i " + filename + " -strict -2 -c:v libx264 -crf 26 -vf scale=480:480 vm.mp4")
+                sendVideoNote("vm.mp4")
                 bot.editMessageText(msgid, "Sending...")
                 f = open("out.mp4", "r")
                 bot.sendVideo(chat_id, f)
@@ -761,6 +755,13 @@ def sendAudioChan(chat_id,file_name,performer,title,caption):
 def sendVoice(chat_id,file_name):
     url = "https://api.telegram.org/bot%s/sendVoice"%(TOKEN)
     files = {'voice': open(file_name, 'rb')}
+    data = {'chat_id' : chat_id}
+    r= requests.post(url, files=files, data=data)
+    print(r.status_code, r.reason, r.content)
+
+def sendVideoNote(chat_id,file_name):
+    url = "https://api.telegram.org/bot%s/sendVideoNote"%(TOKEN)
+    files = {'video_note': open(file_name, 'rb')}
     data = {'chat_id' : chat_id}
     r= requests.post(url, files=files, data=data)
     print(r.status_code, r.reason, r.content)
